fix: restructured the Verifier and holder files
