+++ shared/src/androidMain/kotlin/data/bletransfer/holder/CredentialStore.kt
@@ -1,23 +1,22 @@
 package data.bletransfer.holder
 
 import android.content.Context
-import android.security.identity.IdentityCredentialStore
+import com.android.identity.android.legacy.IdentityCredentialStore
 
 class CredentialStore(
     private val context: Context
 ) {
 
-    fun createIdentityCredentialStore(): IdentityCredentialStore? {
-        return IdentityCredentialStore.getInstance(context)
-//        return if (PreferencesHelper.isHardwareBacked())
-//            IdentityCredentialStore.getHardwareInstance(context)
-//                ?: createKeystoreBackedStore() else createKeystoreBackedStore()
+    fun createIdentityCredentialStore(): IdentityCredentialStore {
+        return if (PreferencesHelper.isHardwareBacked())
+            IdentityCredentialStore.getHardwareInstance(context)
+                ?: createKeystoreBackedStore() else createKeystoreBackedStore()
     }
 
-//    private fun createKeystoreBackedStore(): IdentityCredentialStore {
-//        val keystoreBackedStorageLocation = PreferencesHelper
-//            .getKeystoreBackedStorageLocation(context)
-//        return IdentityCredentialStore
-//            .getKeystoreInstance(context, keystoreBackedStorageLocation)
-//    }
+    private fun createKeystoreBackedStore(): IdentityCredentialStore {
+        val keystoreBackedStorageLocation = PreferencesHelper
+            .getKeystoreBackedStorageLocation(context)
+        return IdentityCredentialStore
+            .getKeystoreInstance(context, keystoreBackedStorageLocation)
+    }
 }
