--- shared/src/androidMain/kotlin/data/holder/transfer/SessionSetup.kt
@@ -1,16 +1,0 @@
-package data.holder.transfer
-
-import com.android.identity.android.legacy.IdentityCredentialStore.CIPHERSUITE_ECDHE_HKDF_ECDSA_WITH_AES_256_GCM_SHA256
-import com.android.identity.android.legacy.PresentationSession
-
-class SessionSetup(
-    private val credentialStore: CredentialStore
-) {
-
-    fun createSession(): PresentationSession {
-        val store = credentialStore.createIdentityCredentialStore()
-        return store.createPresentationSession(
-            CIPHERSUITE_ECDHE_HKDF_ECDSA_WITH_AES_256_GCM_SHA256
-        )
-    }
-}
