--- shared/src/androidMain/kotlin/data/verifier/transfer/TransferStatus.kt
@@ -1,11 +1,0 @@
-package ui.data.transfer
-
-enum class TransferStatus {
-    READER_ENGAGEMENT_READY,
-    ENGAGED,
-    UNUSED,
-    CONNECTED,
-    RESPONSE,
-    DISCONNECTED,
-    ERROR,
-}
