--- shared/src/commonMain/kotlin/data/holder/Holder.kt
@@ -1,14 +1,0 @@
-package data.holder
-
-import androidx.compose.runtime.Composable
-
-expect fun getHolder(): Holder
-
-interface Holder {
-    @Composable
-    fun getRequirements(check: (Boolean) -> Unit)
-
-    fun hold(updateQrCode: (String) -> Unit)
-
-    fun disconnect()
-}
