--- shared/src/iosMain/kotlin/data/holder/Holder.kt
@@ -1,4 +1,0 @@
-package data.holder
-
-
-actual fun getHolder(): Holder { throw NotImplementedError("Holder is not implemented for iOS yet")}
