--- shared/src/iosMain/kotlin/data/verifier/Verifier.kt
@@ -1,3 +1,0 @@
-package data.verifier
-
-actual fun getVerifier(): Verifier { throw NotImplementedError("Verifier is not implemented for iOS yet")}
